Dev lod decoration

+++ Terrain/GUI/GUI.cpp
@@ -303,7 +303,7 @@ void GUI::processKeyboardEvents(CForge::Keyboard* keyboard)
 void GUI::listen(char32_t codepoint)
 {
     //print characters and keycodes for testing purposes
-    std::cout << "Received " << codepoint << std::endl;
+    std::cout << "Received " << (uint32_t)codepoint << std::endl;
 
     //pass the character to the focused widget.
     if (focusedWidget != nullptr) focusedWidget->onKeyPress(codepoint);
